user-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pymongo>=4.10
beanie>=1.24.0
strawberry-graphql[fastapi]>=0.236.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
